feat(Bank.h, bank.cpp) fix bugs and move functions from Bank.h to bank.cpp; also add threads and mutex
